+++ domain-module/src/main/java/domain/store/Store.java
@@ -66,7 +66,7 @@ public class Store {
     @ElementCollection
     @CollectionTable(name = "starredUsers", joinColumns = @JoinColumn(name="storeId", referencedColumnName="storeId"))
     @Cascade(org.hibernate.annotations.CascadeType.ALL)
-    private List<Integer> starredUsers;    // 찜한 유저 아이디들
+    private List<Long> starredUsers;    // 찜한 유저 아이디들
 
     @Column(nullable = false)
     private String crNumber;    // 사업자 등록번호
